requirements.txt
scikit-learn
pandas
joblib
apache-airflow